+++ src/server/api/routers/role.tsx
@@ -3,15 +3,15 @@ import { createTRPCRouter, getProcedure } from "../trpc";
 import { z } from "zod";
 import { RolePayload, accessTypes } from "@/types/prisma";
 import { idSchema } from "@/utils/validation";
-import { DefaultSortOrder } from "@/utils/constants";
+import { defaultSortOrder } from "@/utils/constants";
 
 export const RoleRouter = createTRPCRouter({
-  getRoles: getProcedure(AccessType.readAccess)
+  all: getProcedure(AccessType.readAccess)
     .input(
       z.object({
         search: z.string().trim().default(""),
-        sortOrder: z.enum(["asc", "desc"]).default(DefaultSortOrder),
-      })
+        sortOrder: z.enum(["asc", "desc"]).default(defaultSortOrder),
+      }),
     )
     .query(async ({ input, ctx: { prisma } }) => {
       return prisma.role.findMany({
@@ -26,7 +26,7 @@ export const RoleRouter = createTRPCRouter({
         },
       });
     }),
-  getRole: getProcedure(AccessType.readAccess)
+  byId: getProcedure(AccessType.readAccess)
     .input(z.object({ id: idSchema }))
     .query(async ({ input: { id }, ctx: { prisma } }) => {
       return prisma.role.findUnique({
@@ -34,14 +34,21 @@ export const RoleRouter = createTRPCRouter({
         include: { accesses: true },
       });
     }),
-  createRole: getProcedure(AccessType.createRole)
+  create: getProcedure(AccessType.createRole)
     .input(
       z.object({
         name: z.string(),
         accesses: z.array(z.enum(accessTypes)),
-      })
+      }),
     )
     .mutation(async ({ input: { name, accesses }, ctx: { prisma } }) => {
+      const existingRole = await prisma.role.findFirst({ where: { name } });
+      if (existingRole) {
+        return "Role already exists";
+      }
+      if (accesses.length === 0) {
+        return "Role must have at least one access";
+      }
       return prisma.role.create({
         data: {
           name,
@@ -51,20 +58,31 @@ export const RoleRouter = createTRPCRouter({
         },
       });
     }),
-  deleteRole: getProcedure(AccessType.deleteRole)
+  delete: getProcedure(AccessType.deleteRole)
     .input(z.object({ id: z.string() }))
     .mutation(async ({ input: { id }, ctx: { prisma } }) => {
+      const existingRole = await prisma.role.findFirst({ where: { id } });
+      if (!existingRole) {
+        return "Role not found";
+      }
       return prisma.role.delete({ where: { id } });
     }),
-  addAccessToRole: getProcedure(AccessType.updateRole)
+  addAccess: getProcedure(AccessType.updateRole)
     .input(
       z.object({
         roleId: idSchema,
         accesses: z.array(z.enum(accessTypes)).nonempty(),
-      })
+      }),
     )
     .mutation(
       async ({ input: { roleId, accesses: access }, ctx: { prisma } }) => {
+        const existingRole = await prisma.role.findFirst({
+          where: { id: roleId },
+        });
+        if (!existingRole) {
+          return "Role not found";
+        }
+
         return prisma.role.update({
           where: { id: roleId },
           data: {
@@ -73,14 +91,14 @@ export const RoleRouter = createTRPCRouter({
             },
           },
         });
-      }
+      },
     ),
-  removeAccessFromRole: getProcedure(AccessType.updateRole)
+  removeAccess: getProcedure(AccessType.updateRole)
     .input(
       z.object({
         roleId: idSchema,
         accesses: z.array(z.enum(accessTypes)).nonempty(),
-      })
+      }),
     )
     .mutation(
       async ({ input: { roleId, accesses: access }, ctx: { prisma } }) => {
@@ -92,6 +110,6 @@ export const RoleRouter = createTRPCRouter({
             },
           },
         });
-      }
+      },
     ),
 });
